JK - Removed debugging
